# [READY]  Stop using regex to parse tags files