--- a/cpp/BoostParts/boost/config/stdlib/msl.hpp
+++ b/cpp/BoostParts/boost/config/stdlib/msl.hpp
@@ -75,6 +75,7 @@
 #  define BOOST_NO_CXX11_HDR_ATOMIC
 #  define BOOST_NO_CXX11_STD_ALIGN
 #  define BOOST_NO_CXX11_ADDRESSOF
+#  define BOOST_NO_CXX11_HDR_EXCEPTION
 
 #if defined(__has_include)
 #if !__has_include(<shared_mutex>)