--- a/cpp/BoostParts/boost/config/stdlib/stlport.hpp
+++ b/cpp/BoostParts/boost/config/stdlib/stlport.hpp
@@ -235,6 +235,7 @@ namespace boost { using std::min; using std::max; }
 #  define BOOST_NO_CXX11_HDR_ATOMIC
 #  define BOOST_NO_CXX11_STD_ALIGN
 #  define BOOST_NO_CXX11_ADDRESSOF
+#  define BOOST_NO_CXX11_HDR_EXCEPTION
 
 #if defined(__has_include)
 #if !__has_include(<shared_mutex>)