--- a/cpp/BoostParts/boost/filesystem.hpp
+++ b/cpp/BoostParts/boost/filesystem.hpp
@@ -14,7 +14,10 @@
 
 #  include <boost/filesystem/config.hpp>
 #  include <boost/filesystem/path.hpp>
+#  include <boost/filesystem/exception.hpp>
+#  include <boost/filesystem/directory.hpp>
 #  include <boost/filesystem/operations.hpp>
+#  include <boost/filesystem/file_status.hpp>
 #  include <boost/filesystem/convenience.hpp>
 #  include <boost/filesystem/string_file.hpp>