--- a/cpp/BoostParts/boost/filesystem/path.hpp
+++ b/cpp/BoostParts/boost/filesystem/path.hpp
@@ -27,6 +27,7 @@
 #include <boost/system/error_code.hpp>
 #include <boost/system/system_error.hpp>
 #include <boost/iterator/iterator_facade.hpp>
+#include <boost/iterator/iterator_categories.hpp>
 #include <boost/core/enable_if.hpp>
 #include <boost/io/detail/quoted_manip.hpp>
 #include <boost/functional/hash_fwd.hpp>
@@ -59,6 +60,7 @@ namespace path_detail // intentionally don't use filesystem::detail to not bring
     static BOOST_CONSTEXPR_OR_CONST value_type      dot = Dot;
   };
 
+#if defined(BOOST_NO_CXX17_INLINE_VARIABLES)
   template< typename Char, Char Separator, Char PreferredSeparator, Char Dot >
   BOOST_CONSTEXPR_OR_CONST typename path_constants< Char, Separator, PreferredSeparator, Dot >::value_type
   path_constants< Char, Separator, PreferredSeparator, Dot >::separator;
@@ -68,6 +70,7 @@ namespace path_detail // intentionally don't use filesystem::detail to not bring
   template< typename Char, Char Separator, Char PreferredSeparator, Char Dot >
   BOOST_CONSTEXPR_OR_CONST typename path_constants< Char, Separator, PreferredSeparator, Dot >::value_type
   path_constants< Char, Separator, PreferredSeparator, Dot >::dot;
+#endif
 
 } // namespace path_detail