--- a/cpp/BoostParts/boost/functional/hash.hpp
+++ b/cpp/BoostParts/boost/functional/hash.hpp
@@ -1,6 +1,0 @@
-
-// Copyright 2005-2009 Daniel James.
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#include <boost/container_hash/hash.hpp>