--- a/cpp/BoostParts/boost/preprocessor/seq/detail/is_empty.hpp
+++ b/cpp/BoostParts/boost/preprocessor/seq/detail/is_empty.hpp
@@ -21,29 +21,29 @@
 /* An empty seq is one that is just BOOST_PP_SEQ_NIL */
 #
 # define BOOST_PP_SEQ_DETAIL_IS_EMPTY(seq) \
-	BOOST_PP_COMPL \
-		( \
-		BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY(seq) \
-		) \
+    BOOST_PP_COMPL \
+        ( \
+        BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY(seq) \
+        ) \
 /**/
 #
 # define BOOST_PP_SEQ_DETAIL_IS_EMPTY_SIZE(size) \
-	BOOST_PP_COMPL \
-		( \
-		BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY_SIZE(size) \
-		) \
+    BOOST_PP_COMPL \
+        ( \
+        BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY_SIZE(size) \
+        ) \
 /**/
 #
 # define BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY(seq) \
-	BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY_SIZE(BOOST_PP_SEQ_DETAIL_EMPTY_SIZE(seq)) \
+    BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY_SIZE(BOOST_PP_SEQ_DETAIL_EMPTY_SIZE(seq)) \
 /**/
 #
 # define BOOST_PP_SEQ_DETAIL_IS_NOT_EMPTY_SIZE(size) \
-	BOOST_PP_BOOL(size) \
+    BOOST_PP_BOOL(size) \
 /**/
 #
 # define BOOST_PP_SEQ_DETAIL_EMPTY_SIZE(seq) \
-	BOOST_PP_DEC(BOOST_PP_SEQ_SIZE(seq (nil))) \
+    BOOST_PP_DEC(BOOST_PP_SEQ_SIZE(seq (nil))) \
 /**/
 #
 # endif