@bstaletic 2020-03-09
`string_view` is C++17. Yes, we can improve this code with `string_view`, but not until we move to C++17.
We could grab a `const auto`, but since `lines` is already `const`, so `auto&&` is already `const`-correct.